check if user is admin for users management

+++ src/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Put, Delete, Param, Body, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Put, Delete, Param, Body, UseGuards, Request, ForbiddenException } from '@nestjs/common';
 import { User } from './user.model';
 import { UsersService } from './users.service';
 import { JwtAuthGuard } from './../auth/jwt-auth.guard';
@@ -9,35 +9,50 @@ export class UsersController {
 
     @UseGuards(JwtAuthGuard)
     @Get()
-    async getUsers(): Promise<User[]> {
+    async getUsers(@Request() req): Promise<User[]> {
+        const { isAdmin } = await this.usersService.getUser(req.user.username)
+        if (!isAdmin) throw new ForbiddenException()
+
         const users = await this.usersService.getUsers()
         return users
     }
 
     @UseGuards(JwtAuthGuard)
     @Get(':id')
-    async getUser(@Param('id') id: string): Promise<User> {
+    async getUser(@Param('id') id: string, @Request() req): Promise<User> {
+        const { isAdmin } = await this.usersService.getUser(req.user.username)
+        if (!isAdmin) throw new ForbiddenException()
+
         const user = await this.usersService.getUser(id)
         return user
     }
 
     @UseGuards(JwtAuthGuard)
     @Post('')
-    async createUser(@Body() newUser: User): Promise<User> {
+    async createUser(@Body() newUser: User, @Request() req): Promise<User> {
+        const { isAdmin } = await this.usersService.getUser(req.user.username)
+        if (!isAdmin) throw new ForbiddenException()
+
         const insertedUser = await this.usersService.insertUser(newUser)
         return insertedUser
     }
 
     @UseGuards(JwtAuthGuard)
     @Put(':id')
-    async updateUser(@Param('id') id: string, @Body() user: User): Promise<User> {
+    async updateUser(@Param('id') id: string, @Body() user: User, @Request() req): Promise<User> {
+        const { isAdmin } = await this.usersService.getUser(req.user.username)
+        if (!isAdmin) throw new ForbiddenException()
+
         const updatedUser = await this.usersService.updateUser(id, user)
         return updatedUser
     }
 
     @UseGuards(JwtAuthGuard)
     @Delete(':id')
-    async deleteUser(@Param('id') id: string): Promise<User> {
+    async deleteUser(@Param('id') id: string, @Request() req): Promise<User> {
+        const { isAdmin } = await this.usersService.getUser(req.user.username)
+        if (!isAdmin) throw new ForbiddenException()
+
         const deletedUser = await this.usersService.deleteUser(id)
         return deletedUser
     }
